fix border position calculations
For one, this fixes a bug where the x and y direction of the aspect
ratio calculation were switched, leading to incorrect placement near the
corners of the frame. When using offsets from the actual frame border,
the aspect ratio of the offset border is now used for correct behavior
in this case.

re #13